Creation of mailtm.js api wapper
This is a nodejs mailtm api wrapper. Very basic and easy to use.  Each function responds back with the entire response with a success boolean added except get_all_messages just responds with the message array. Also note that you can only send 8 requests per second so there is code for the get_all_messages as well but this only ensures that it won't go above limit for the calling of that function. All other management to ensure you don't pass the 8 requests a second must be done by you. I am not responsible for any misuse.